Refactor APIs to work on Floats instead of Ints.
This applies to the Canvas object, the Scene API, the Viewport, and the
geometry. The reasoning is that these are used for
representing/rendering world objects, which should not be in pixel
coordinates, because they can't (due to physics simulation and
precision).

Canvas can, on top of that, be scaled and rotated, which will make its
integer-based API insufficient. The commit contains an example in the
hello project, which shows what can be done now that would have not
worked very well before (leading to square of uneven sizes).