Support for input event processors.
This change is needed to accomodate browser contraints for the
html5 backend. Some browsers will only play music or open links
in response to a user input. The current processing of events happen
in an async way, with the event being detected and then added to a
queue, and only processed by the game's code on the next renderering
call. By the time we process and act on it (play a sound, or open
a link), the browser lost track that it was user-initiated, and will
prevent the action. Note that this is mostly true for Safari, both
Chrome and Firefox seem fine with the previous behavior, but this change
should makes things more portable.